Allow empty bindings in override bindings
This allows override bindings to not declare a bindings array. This can
be useful if the bindings menu will be populated with other overrides.
For instance, the sample config will now work:

```json
{
  "whichkey.bindingOverrides": [
    {
      "keys": "g.H",
      "name": "Hunks",
      "type": "bindings"
    },
    {
      "keys": "g.H.s",
      "name": "Stage hunk",
      "type": "command",
      "command": "git.stageSelectedRanges"
    }
  ]
}
```

Previously, the former config would fail to populate the "Hunks" menu
due to the missing bindings entry defaulting to null instead of an empty
array. This would cause the overrideBindings function to behave as if
the "Hunks" menu didn't exist, and fail to append the item.